requirements.txt
flask
openai
python-dotenv
moviepy
numpy